Move utils file to relevant location.
